Updated Solution - GfG to GitHub

+++ Difficulty: Medium/Distance from the Source (Bellman-Ford Algorithm)/distance-from-the-source-bellman-ford-algorithm.cpp
@@ -13,7 +13,7 @@ class Solution {
     *   V: number of vertices
     */
     vector<int> bellman_ford(int V, vector<vector<int>>& edges, int S) {
-        vector<int> dist(V, (int)1e8);
+        vector<int> dist(V, 1e8);
         dist[S] = 0;
         for(int i = 0; i < V-1; i++){
             for(auto it : edges){
